Update policykit changes for NetworkManager for Trusty.
We need to allow org.freedesktop.NetworkManager.settings.modify.own
for lightdm-user so that networks can be connected to on the
login screen.  I also think that giving lightdm the permission
to org.freedesktop.NetworkManager.settings.modify.system should
be useful so that networks made in the login screen can be
shared and their passwords saved.  It also appears that the
"enable-disable-network;enable-disable-wifi" stuff is no longer needed.