Do not install flashblock anymore.
Flashblock needs an update to newer version to work at all, but even
then whitelisting is broken.  Flashblock causes issues on some sites, and
whitelisting is necessary for those (there is no apperent flashblock-icon
for user to click).  It also blocks some content which it should not
(such as links to plain mp3-files), or at least has done that in some
older versions).  It also appears that major Finnish news outlets no
longer use Flash for advertising, so need for Flashblock should be
rather low.